Renamed removeTask to taskAboutToBeRemoved.
The purpose of ThreadedApplication::AbstractListener::removeTask is
to enable classes which cannot derive from ThreadedApplication to be
notified once a task is about to be removed. As such, the name
removeTask was a bit misleading and has thus been changed to
taskAboutToBeRemoved.
As the browser makes use of this method, the corresponding files have
been adapted accordingly.